+++ parsers/erlang.c
@@ -66,7 +66,7 @@ static const unsigned char *parseIdentifier (
 	vStringClear (identifier);
 	while (isIdentifierCharacter ((int) *cp))
 	{
-		vStringPut (identifier, (int) *cp);
+		vStringPut (identifier, *cp);
 		++cp;
 	}
 	return cp;
