+++ parsers/scheme.c
@@ -49,7 +49,7 @@ static void readIdentifier (vString *const name, const unsigned char *cp)
 	vStringClear (name);
 	/* Go till you get to white space or a syntactic break */
 	for (p = cp; *p != '\0' && *p != '(' && *p != ')' && !isspace (*p); p++)
-		vStringPut (name, (int) *p);
+		vStringPut (name, *p);
 }
 
 static void findSchemeTags (void)
